+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2
 
 - Added option to modify xAxis height and yAxis width
+- Added option to specify Axis label font size
 
 ## 0.0.1