+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added option to modify xAxis height and yAxis width
 - Added option to specify Axis label font size
+- Exported calculateTextSize function
+- Improve point detection logic
 
 ## 0.0.1